Problem: database flag and orbs arguments were confusing, how can we map multiple orbs to a single database?
Solution: Drop the arguments and database flag and rely only on the
directory until we better understand what is the best way to override
the defaults.
The defaults as implemented here will use the current orb as the first
directory below the workspace on the current directory tree.
When the current workspace is the current directory, or the current
directory is outside the workspace tree then we consider all orbs listed
in the config.

+++ cmd/migrate.go
@@ -5,21 +5,22 @@ import (
 	"database/sql"
 	"database/sql/driver"
 	"encoding/json"
-	"errors"
 	"fmt"
+	"maps"
+	"os"
+	"path"
+	"path/filepath"
+	"strings"
+
 	"github.com/charmbracelet/log"
 	"github.com/lib/pq"
 	"github.com/omnigres/cli/orb"
 	"github.com/samber/lo"
 	"github.com/spf13/cobra"
-	"maps"
-	"os"
-	"path"
-	"strings"
 )
 
 var migrateCmd = &cobra.Command{
-	Use:   "migrate [orb...]",
+	Use:   "migrate",
 	Short: "Migrate orbs",
 	Long:  `By default, will migrate all listed orbs`,
 	Run: func(cmd *cobra.Command, args []string) {
@@ -29,18 +30,61 @@ var migrateCmd = &cobra.Command{
 		if err != nil {
 			log.Fatal(err)
 		}
-		ctx := context.Background()
-		orbs := lo.Map(cluster.Config().Orbs, func(cfg orb.OrbCfg, _ int) string { return cfg.Name })
-		err = migrate(ctx, cluster, dbReset, orbs, orbs)
 
+		log.Debug("Workspace", "workspace", workspace)
+		cwd, err := os.Getwd()
+		if err != nil {
+			log.Fatal(err)
+		}
+
+		orbs, err := currentOrbs(cluster, cwd)
+		if err != nil {
+			log.Fatal(err)
+		}
+
+		ctx := context.Background()
+		log.Debug("Migrating orbs", "orbs", orbs)
+		err = assembleOrbs(
+			ctx,
+			cluster,
+			dbReset,
+			orbs,
+			func(orbName string) string { return orbName },
+		)
 	},
 }
 
-func migrate(ctx context.Context, cluster orb.OrbCluster, dbReset bool, orbs []string, databases []string) (err error) {
-	if len(orbs) != len(databases) {
-		err = errors.New("orbs and databases have to be of the same size")
-		return
+func currentOrbs(cluster orb.OrbCluster, dir string) ([]string, error) {
+	if dir == workspace {
+		// return all orbs when running in workspace root
+		return lo.Map(
+			cluster.Config().Orbs,
+			func(cfg orb.OrbCfg, _ int) string { return cfg.Name },
+		), nil
 	}
+
+	parent := filepath.Dir(dir)
+	if parent == workspace {
+		// First level below workspace, this is the current orb
+		cwd, err := os.Getwd()
+		return []string{filepath.Base(cwd)}, err
+	}
+
+	if parent == dir {
+		// Reached the root, return as in workspace root
+		return currentOrbs(cluster, workspace)
+	}
+
+	return currentOrbs(cluster, parent)
+}
+
+func assembleOrbs(
+	ctx context.Context,
+	cluster orb.OrbCluster,
+	dbReset bool,
+	orbs []string,
+	databaseForOrb func(string) string,
+) (err error) {
 	logger := log.New(os.Stdout)
 	logger.SetReportTimestamp(true)
 	logger.Info("Starting migration...")
@@ -50,33 +94,29 @@ func migrate(ctx context.Context, cluster orb.OrbCluster, dbReset bool, orbs []s
 	if err != nil {
 		return
 	}
-	for i, dbName := range databases {
+	for _, orbName := range orbs {
+		dbName := databaseForOrb(orbName)
 		logger.SetPrefix(fmt.Sprintf("[%s] ", dbName))
-		var datname string
-		if databaseName != "" {
-			dbName = databaseName
-		}
-		err := db.QueryRowContext(ctx, `select datname from pg_database where datname = $1`, dbName).Scan(&datname)
-		if err != nil && err != sql.ErrNoRows {
+		var dbExists bool
+		err := db.QueryRowContext(
+			ctx,
+			`select exists(select from pg_database where datname = $1)`,
+			dbName,
+		).Scan(&dbExists)
+		if err != nil {
 			log.Fatal(err)
 		}
 
-		createDb := false
-
-		if err == sql.ErrNoRows {
-			createDb = true
-		}
-
-		if dbReset {
+		if dbReset && dbExists {
 			if err != sql.ErrNoRows {
 				_, err = db.ExecContext(ctx, fmt.Sprintf(`drop database %q`, dbName))
 				if err != nil {
 					log.Fatal(err)
 				}
-				createDb = true
 			}
 		}
-		if createDb {
+
+		if dbReset || !dbExists {
 			_, err = db.ExecContext(ctx, fmt.Sprintf(`create database %q`, dbName))
 			if err != nil {
 				log.Fatal(err)
@@ -117,10 +157,7 @@ func migrate(ctx context.Context, cluster orb.OrbCluster, dbReset bool, orbs []s
 			return nil
 		})
 
-		orbSource := path.Join(orbs[i], "src")
-		if orbs[i] == "." {
-			orbSource = ""
-		}
+		orbSource := path.Join(dbName, "src")
 		rows, err := conn.QueryContext(ctx,
 			`select migration_filename, migration_statement, execution_error from omni_schema.assemble_schema($1, omni_vfs.local_fs('/mnt/host'), $2) where execution_error is not null`,
 			fmt.Sprintf("dbname=%s user=omnigres", dbName), orbSource)
@@ -143,10 +180,8 @@ func migrate(ctx context.Context, cluster orb.OrbCluster, dbReset bool, orbs []s
 }
 
 var dbReset bool
-var databaseName string
 
 func init() {
 	rootCmd.AddCommand(migrateCmd)
-	migrateCmd.Flags().StringVarP(&databaseName, "db", "d", "", "database name")
 	migrateCmd.Flags().BoolVarP(&dbReset, "dbReset", "r", false, "dbReset")
 }
